--- akka-http-tests-java8/src/test/resources/reference.conf
@@ -1,6 +1,0 @@
-akka {
-  actor {
-    serialize-creators = on
-    serialize-messages = on
-  }
-}
