+++ akka-http-tests/src/test/java/docs/http/javadsl/server/HandlerExampleDocTest.java
